requirements.txt
ultralytics 
torch 
opencv-python 
pyserial 
pyyaml 
requests
torchvision